Lutron:  Added Support for Phantom Buttons 1-100
Moved functionality to Bridge thing instead of Phantom Button thing.
